Convert Enabled into a Boolean
This one was weird. Several of the properties that Get-NetFirewallRule emits
have their own custom type. To convert it into a Boolean, you first have to
cast it to a String.

This will only work properly in 19.1.6.0 or later. We recently fixed some
object recognition issues. It still works in older versions, but Enabled
is treated as a String.